use size_t instead of uint16_t for bytes

+++ libtransmission/peer-common.h
@@ -265,8 +265,8 @@ struct tr_peer
     // how many requests we made to this peer and then canceled
     tr_recentHistory<uint16_t> cancels_sent_to_peer;
 
-    tr_recentHistory<uint16_t> bytes_sent_to_client;
-    tr_recentHistory<uint16_t> bytes_sent_to_peer;
+    tr_recentHistory<size_t> bytes_sent_to_client;
+    tr_recentHistory<size_t> bytes_sent_to_peer;
 };
 
 // ---
